Move code into app folder
