+++ max-mcp-openai-agents-sdk/.gitignore
@@ -1,0 +1,6 @@
+.web
+*.py[cod]
+assets/external/
+__pycache__/
+.states
+*.db
