+++ max-mcp-openai-agents-sdk/rxconfig.py
@@ -1,0 +1,5 @@
+import reflex as rx
+
+config = rx.Config(
+    app_name="app",
+)
